First Draft - Arranger router updates
