+++ api/index.js
@@ -8,4 +8,4 @@ require('@babel/register')({
   ],
 });
 
-require('./src/index');
+require('./src/index.ts');
