+++ api/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@hcmi-portal/api",
   "version": "0.0.1",
-  "description": "``` npm i @overture-stack/arranger-server ```",
+  "description": "``` npm i @overture-stack/arranger-graphql-router ```",
   "private": true,
   "main": "index.js",
   "scripts": {
@@ -11,7 +11,9 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@overture-stack/arranger-server": "^3.0.4",
+    "@overture-stack/arranger-types": "1.0.0-rc.2",
+    "@overture-stack/arranger-graphql-router": "1.0.0-rc.2",
+    "@overture-stack/sqon-builder": "^1.1.0",
     "JSONStream": "^1.3.5",
     "babel-polyfill": "^6.26.0",
     "dotenv": "^5.0.1",
