+++ elasticsearch/arranger_metadata/base.json
@@ -1,4 +1,4 @@
 {
   "documentType": "model",
-  "index": "hcmi"
+  "esIndex": "hcmi"
 }
